Move Topaz denoising service to on which can run as a container on GPU nodes (#72)
