+++ Dockerfiles/topaz
@@ -1,0 +1,19 @@
+FROM python:3.10
+
+ENV VIRTUAL_ENV=/venv
+RUN python3 -m venv $VIRTUAL_ENV
+ENV PATH="$VIRTUAL_ENV/bin:$PATH"
+ARG groupid
+ARG userid
+ARG groupname
+
+# Install cryoem-services
+COPY --chown="${userid}":"${groupid}" packages/cryoem-services ./cryoem-services
+RUN python -m pip install ./cryoem-services
+
+# Install topaz
+RUN python -m pip install topaz-em
+
+# Create EM user
+RUN groupadd -r -g "${groupid}" "${groupname}" && useradd -r -M "${groupname}" -u "${userid}" -g "${groupname}"
+USER "${userid}":"${groupid}"
