+++ Helm/charts/denoise/templates/deployment.yaml
@@ -1,23 +1,23 @@
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: tomo-align
+  name: denoise
   namespace: {{ .Values.global.namespace }}
 spec:
   replicas: {{ .Values.replicas }}
   selector:
     matchLabels:
-      app: tomo-align
+      app: denoise
   template:
     metadata:
       labels:
-        app: tomo-align
+        app: denoise
     spec:
       securityContext:
         runAsUser: {{ .Values.global.runAsUser }}
         runAsGroup: {{ .Values.global.runAsGroup }}
       containers:
-      - name: tomo-align-runner
+      - name: denoise-runner
         image: {{ .Values.image }}
         imagePullPolicy: Always
         resources:
@@ -38,10 +38,6 @@ spec:
           value: "/zocalo/config/{{ .Values.global.configFileName }}"
         - name: ISPYB_CREDENTIALS
           value: "/zocalo/secrets/{{ .Values.global.dbFileName }}"
-        - name: LD_LIBRARY_PATH
-          value: "/IMOD/4.11.1/IMOD/lib:/IMOD/4.11.1/IMOD/qtlib"
-        - name: IMOD_DIR
-          value: "/IMOD/4.11.1/IMOD"
         volumeMounts:
         - name: config-file
           mountPath: /zocalo/config
@@ -58,7 +54,7 @@ spec:
         projected:
           defaultMode: 0444
           sources:
-          - secret: 
+          - secret:
               name: {{ .Values.global.dbSecretName }}
           - secret:
               name: {{ .Values.global.rmqSecretName }}
@@ -88,16 +84,16 @@ spec:
 apiVersion: keda.sh/v1alpha1
 kind: ScaledObject
 metadata:
-  name: tomo-align
+  name: denoise
   namespace: {{ .Values.global.namespace }}
 spec:
   scaleTargetRef:
-    name: tomo-align
+    name: denoise
   triggers:
     - type: rabbitmq
       metadata:
         host: {{ .Values.global.rmqHost }}
-        queueName: tomo_align
+        queueName: denoise
         mode: QueueLength
         value: "{{ .Values.queueLengthTrigger }}"
   minReplicaCount: {{ .Values.minReplicaCount }}
