+++ Helm/charts/denoise_slurm/Chart.yaml
@@ -1,4 +1,4 @@
 apiVersion: v2
 name: denoise_slurm
-description: A Helm chart for EM Denoise service
+description: A Helm chart for EM Denoise service which submits jobs to the IRIS cluster
 version: 0.5.3
